Add files via upload
One sensor mounted on top of one servo.
At the start the sensor detects how far the wall is, which will determine the  hovercraft’s motion.
As soon as that reading is taken, the sensor will rotate to the right to search for holes in the wall. This can be done every 0.5 meters or something.
Since we know through the equations of motion where the hovercraft will be at each moment in time so sensor measurements directed straight won’t be needed as frequently.
As the hovercraft moves towards the end of the straight it will detect a hole and would then enter a corner/turning process. And then repeat the procedure for straight line operation.
My idea was that we program the servo to point the sensor in specific directions based on our knowledge of the track. Basically the pattern is 2 right turns followed by 2 left turns. But I don’t know if programming it this way would be cheating.
In that case we can see if the servo/laser combination is fast enough to alternate between detecting holes on the right and left.