Create Simulation 1st attempt
Added a loop to make it easy to feed the simulation instructions (for motion).
Made the yaw torque a function of surge torque (trigonometry) as would be the case in real life (i think)

Next I will add a mechanic which will describe how the sway behaves under rudder rotation. For this I'm planning on describing a new frame of reference (inertial). That way I can see exactly how the body is moving and then assign the sway vectors.